Realistic load

+++ java/src/main/java/org/wooque/bench/Bench.java
@@ -20,15 +20,14 @@ public class Bench extends AbstractVerticle {
         db = JDBCClient.createShared(vertx, new JsonObject()
                 .put("url", "jdbc:postgresql://127.0.0.1/bench?user=bench&password=bench")
                 .put("driver_class", "org.postgresql.Driver")
-                .put("max_pool_size", 128));
+                .put("max_pool_size", 32));
 
         vertx.createHttpServer().requestHandler(r -> db.getConnection(conn -> {
             final SQLConnection connection = conn.result();
 
-            int coin = ThreadLocalRandom.current().nextInt(0, 4);
+            int coin = ThreadLocalRandom.current().nextInt(0, 10);
             String data;
-            switch (coin) {
-                case 0:
+            if (coin < 5) {
                     connection.query("SELECT txt FROM tst LIMIT 1", rs -> {
                         List rows = rs.result().getRows();
                         String resp;
@@ -40,8 +39,8 @@ public class Bench extends AbstractVerticle {
                         connection.close();
                         r.response().end(resp);
                     });
-                    break;
-                case 1:
+            
+            } else if (coin < 7) {
                     data = UUID.randomUUID().toString();
                     connection.updateWithParams("INSERT INTO tst(txt) VALUES (?)", new JsonArray().add(data), rs -> {
                         rs.result();
@@ -49,8 +48,8 @@ public class Bench extends AbstractVerticle {
                         String resp = new JsonObject().put("txt", data).toString();
                         r.response().end(resp);
                     });
-                    break;
-                case 2:
+            
+            } else if (coin < 9) {
                     data = UUID.randomUUID().toString();
                     connection.updateWithParams("UPDATE tst SET txt=? WHERE id=(SELECT id FROM tst LIMIT 1)", new JsonArray().add(data), rs -> {
                         rs.result();
@@ -58,8 +57,8 @@ public class Bench extends AbstractVerticle {
                         String resp = new JsonObject().put("txt", data).toString();
                         r.response().end(resp);
                     });
-                    break;
-                case 3:
+            
+            } else if (coin == 9) {
                     connection.query("DELETE FROM tst WHERE id=(SELECT id FROM tst LIMIT 1) RETURNING id", rs -> {
                         List rows = rs.result().getRows();
                         String resp;
@@ -71,8 +70,8 @@ public class Bench extends AbstractVerticle {
                         connection.close();
                         r.response().end(resp);
                     });
-                    break;
-                default:
+            
+            } else {
                     connection.close();
                     data = new JsonObject().put("error", "Wrong coin").toString();
                     r.response().end(data);
