+++ java_undertow/src/main/java/org/wooque/bench/Bench.java
@@ -20,7 +20,7 @@ public class Bench {
     public static void main(final String[] args) {
         final HikariDataSource ds = new HikariDataSource();
         ds.setJdbcUrl("jdbc:postgresql://127.0.0.1:5432/bench");
-        ds.setMaximumPoolSize(128);
+        ds.setMaximumPoolSize(32);
         ds.setUsername("bench");
         ds.setPassword("bench");
 
@@ -34,13 +34,12 @@ public class Bench {
                             return;
                         }
                         final Connection conn = ds.getConnection();
-                        int coin = ThreadLocalRandom.current().nextInt(0, 4);
+                        int coin = ThreadLocalRandom.current().nextInt(0, 10);
                         String data;
                         PreparedStatement stmt;
                         ResultSet rs;
                         HashMap<String, String> resp;
-                        switch (coin) {
-                            case 0:
+                        if (coin < 5) {
                                 stmt = conn.prepareStatement("SELECT txt FROM tst LIMIT 1");
                                 rs = stmt.executeQuery();
                                 if (rs.next()) {
@@ -50,24 +49,24 @@ public class Bench {
                                 }
                                 resp = new HashMap<>();
                                 resp.put("txt", data);
-                                break;
-                            case 1:
+
+                        } else if (coin < 7) {
                                 data = UUID.randomUUID().toString();
                                 stmt = conn.prepareStatement("INSERT INTO tst(txt) VALUES (?)");
                                 stmt.setString(1, data);
                                 stmt.executeUpdate();
                                 resp = new HashMap<>();
                                 resp.put("txt", data);
-                                break;
-                            case 2:
+
+                        } else if (coin < 9) {
                                 data = UUID.randomUUID().toString();
                                 stmt = conn.prepareStatement("UPDATE tst SET txt=? WHERE id=(SELECT id FROM tst LIMIT 1)");
                                 stmt.setString(1, data);
                                 stmt.executeUpdate();
                                 resp = new HashMap<>();
                                 resp.put("txt", data);
-                                break;
-                            case 3:
+
+                        } else if (coin == 9) {
                                 stmt = conn.prepareStatement("DELETE FROM tst WHERE id=(SELECT id FROM tst LIMIT 1) RETURNING id");
                                 rs = stmt.executeQuery();
                                 if (rs.next()) {
@@ -77,8 +76,8 @@ public class Bench {
                                 }
                                 resp = new HashMap<>();
                                 resp.put("id", data);
-                                break;
-                            default:
+                        
+                        } else {
                                 resp = new HashMap<>();
                                 resp.put("error", "Wrong coin");
                         }
@@ -89,7 +88,7 @@ public class Bench {
                         exchange.getResponseSender().send(data);
                     }
                 })
-                .setWorkerThreads(200)
+                .setWorkerThreads(128)
                 .build();
         server.start();
     }
